refactor: modularize backend
